refactor: 로또 제한 사이즈 상수 LottoNumbers에서만 관리
