+++ src/main/java/lotto/controller/LottoController.java
@@ -7,8 +7,8 @@ import lotto.model.lotto.*;
 import lotto.model.result.Money;
 import lotto.model.result.RateOfReturn;
 import lotto.model.result.WinningResult;
-import lotto.model.winningnumber.WinningLotto;
-import lotto.model.winningnumber.WinningLottoResponse;
+import lotto.model.lotto.WinningLotto;
+import lotto.model.lotto.WinningLottoResponse;
 import lotto.utils.ConverterUtils;
 import lotto.view.InputView;
 import lotto.view.OutputView;
